Remove class traits
